+++ src/index.js
@@ -5,7 +5,7 @@
 
 export { initGraph, loadGraph, saveGraph } from './graph.js';
 export { createEdge, queryEdges, removeEdge, EDGE_TYPES } from './edges.js';
-export { getNodes, hasNode, getNode, getNodesByPrefix } from './nodes.js';
+export { getNodes, hasNode, getNode, getNodesByPrefix, setNodeProperty, unsetNodeProperty } from './nodes.js';
 export { computeStatus } from './status.js';
 export { importFile, importData, parseImportFile, validateImportData } from './import.js';
 export { importFromMarkdown, parseFrontmatter } from './frontmatter.js';
@@ -16,7 +16,7 @@ export {
   NODE_ID_REGEX, NODE_ID_MAX_LENGTH, CANONICAL_PREFIXES, SYSTEM_PREFIXES, ALL_PREFIXES,
   LOW_CONFIDENCE_THRESHOLD,
 } from './validators.js';
-export { defineView, declareView, renderView, listViews, resetViews } from './views.js';
+export { defineView, declareView, renderView, listViews, resetViews, classifyStatus } from './views.js';
 export { parseDirectives, processCommit } from './hooks.js';
 export {
   getCurrentTick, recordEpoch, lookupEpoch,